+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ### [0.4.0] - Unreleased
 - Addded test for error on extra fields.
+- Added typing information to IsccMeta.iscc_obj property
 - Updated definition of `meta`-field
 
 ### [0.3.9] - 2022-07-03